Better client server sync
So long as both client and host have a directory next to their
executable named Card_Sprites that contains png images for each card. If
a client is missing images or their images are named incorrectly a
generic missing data image will show for the front of the card.

Images all get resized to card size. (1050 x 750 px) for now.

The side of the card syncs correctly on client connection for all cards
now.