Added a simple ui hand system
Users can now add cards to their hands by pressing "a" when hovering
over a card. If they do that the network object for the card gets
deleted and a cliet side ui object gets made for that player. The player
can drag the card from out their hand to drop it back into play.